tx-submit: accept byron txs in the right binary format
This is sadly a little confusing. In Byron there are 4 kinds of on-chain
objects, and only one of those is normal txs. So we have a type called
GenTx ByronBlock that is the "generalised tx" that covers all four
cases. This is the type that is sent between mempools.

But the actual on-chain Byron tx type is a TxAux. The tx-submit API
was using the GenTx format. The cardano-cli tool was doing the same
thing. While this is self consistent this is also very unhelpful, since
we should use the real format as it exists on the chain, not the
intermediate wrapper format.

Doing it this way means we are consistent with client side libraries,
and we are using a format where they correspond exactly to the on-chain
format and we can decode and print them, check their hashes etc.

The code isn't perfect. We can do better by handling this GenTx vs Tx
distinction in a more general way so that we can write the CLI code more
generically between Byron, Shelley and mock protocols & ledgers.

In addition, this code temporarily relies on a couple new functions that
belong in libraries lower down the stack. These local copies can be
deleted as soon as the dependencies are updated.