Refactoring our use of SqlBackend.
There are lots of places in this code where we run in a `ReaderT
SqlBackend` monad *and* pass a `SqlBackend` around manually.

I've neatened this up as much as I can by removing the
manual backend-passing, and hoisting the `runQuery backend` call into
Servant's hoistServer mechanism.

The end result is, IMHO, much cleaner.